Made nonuniform-struct-constructors macro.
This is so we can wrap either a pointer to a struct or a newly
allocated blob, which is needed in cases where you must pass in
a pointer to a struct that you allocated yourself.

It also allows using #f for cases where you can pass a NULL pointer.